secure user endpoint

+++ routers/user.py
@@ -1,11 +1,13 @@
 from typing import Annotated
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
-from auth import Hash, oauth2_scheme, get_current_user
+
+from auth import Hash, get_current_user
+from auth.auth import is_admin
 from services.db_service import get_db
 from models import User
 from schemas import UserRequest, UserResponse
-from helpers.exceptions import http_exception_not_found, http_exception_conflict
+from helpers.exceptions import http_exception_not_found, http_exception_conflict, http_exception_forbidden
 
 router = APIRouter(tags=["user"], prefix="/users")
 
@@ -17,7 +19,7 @@ async def get_all_users(db: Annotated[Session, Depends(get_db)]):
 
 
 @router.get("/me", response_model=UserResponse)
-async def read_users_me(current_user: Annotated[User, Depends(get_current_user)]):
+async def get_my_user(current_user: Annotated[User, Depends(get_current_user)]):
     return current_user
 
 
@@ -61,13 +63,15 @@ async def update_user(
         db: Annotated[Session, Depends(get_db)],
         current_user: Annotated[User, Depends(get_current_user)]
 ):
-    db_user = db.query(User).filter(User.id == user_id).first()
+    db_user: User = db.query(User).filter(User.id == user_id).first()
     if not db_user:
         raise http_exception_not_found(f"User with id {user_id} not found")
+    if db_user.id != current_user.id and not is_admin(current_user):
+        raise http_exception_forbidden()
 
-    if db_user.username == request.username:
+    if db.query(User).filter(User.username == request.username).first() and db_user.username != request.username:
         raise http_exception_conflict("User with this username already exists")
-    if db_user.email == request.email:
+    if db.query(User).filter(User.email == request.email).first() and db_user.email != request.email:
         raise http_exception_conflict("User with this email already exists")
 
     request.password = Hash.bcrypt(request.password)
@@ -81,10 +85,17 @@ async def update_user(
 
 
 @router.delete("/delete/{user_id}")
-async def delete_user(user_id: int, db: Annotated[Session, Depends(get_db)]):
+async def delete_user(
+        user_id: int,
+        db: Annotated[Session, Depends(get_db)],
+        current_user: Annotated[User,
+        Depends(get_current_user)]
+):
     db_user = db.query(User).filter(User.id == user_id).first()
     if not db_user:
         raise http_exception_not_found(f"User with id {user_id} not found")
+    if db_user.id != current_user.id and not is_admin(current_user):
+        raise http_exception_forbidden()
 
     db.delete(db_user)
     db.commit()
